feat(TypeHandlerLibrary): move `ConstructorLibrary`, `CollectionTypeHandlerFactory`, `OjbectFiledMapTypeHandlerFactory` and they tests  to `TypeHandlerWidget` subsystem.
