+++ subsystems/TypeHandlerLibrary/src/test/java/org/terasology/persistence/typeHandling/TypeHandlerLibraryTest.java
@@ -5,13 +5,19 @@ package org.terasology.persistence.typeHandling;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.reflections.Reflections;
+import org.terasology.persistence.typeHandling.coreTypes.CollectionTypeHandler;
 import org.terasology.persistence.typeHandling.coreTypes.EnumTypeHandler;
+import org.terasology.persistence.typeHandling.coreTypes.ObjectFieldMapTypeHandler;
 import org.terasology.persistence.typeHandling.coreTypes.RuntimeDelegatingTypeHandler;
 import org.terasology.persistence.typeHandling.coreTypes.StringMapTypeHandler;
+import org.terasology.reflection.MappedContainer;
 import org.terasology.reflection.TypeInfo;
 
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Queue;
+import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -25,6 +31,34 @@ class TypeHandlerLibraryTest {
         reflections = new Reflections(TypeHandlerLibraryTest.class.getClassLoader());
         typeHandlerLibrary = new TypeHandlerLibrary(reflections);
         TypeHandlerLibrary.populateBuiltInHandlers(typeHandlerLibrary);
+    }
+
+    @MappedContainer
+    private static class AMappedContainer {}
+
+    @Test
+    public void testMappedContainerHandler() {
+        TypeHandler<AMappedContainer> handler = typeHandlerLibrary.getTypeHandler(AMappedContainer.class).get();
+
+        assertTrue(handler instanceof ObjectFieldMapTypeHandler);
+    }
+
+    @Test
+    public void testCollectionHandler() {
+        TypeHandler<Set<Integer>> setHandler =
+                typeHandlerLibrary.getTypeHandler(new TypeInfo<Set<Integer>>() {}).get();
+
+        assertTrue(setHandler instanceof CollectionTypeHandler);
+
+        TypeHandler<List<Integer>> listHandler =
+                typeHandlerLibrary.getTypeHandler(new TypeInfo<List<Integer>>() {}).get();
+
+        assertTrue(listHandler instanceof CollectionTypeHandler);
+
+        TypeHandler<Queue<Integer>> queueHandler =
+                typeHandlerLibrary.getTypeHandler(new TypeInfo<Queue<Integer>>() {}).get();
+
+        assertTrue(queueHandler instanceof CollectionTypeHandler);
     }
 
     @Test
